+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [1.2.3] - 2024-10-27
+- Force save project before merge
+- Show all not existing profiles
+- Add exclude action for not existing profiles
+
 ## [1.2.2] - 2024-10-08
 - Minor changes
 
@@ -22,7 +27,8 @@
 ## [1.0.2] - 2024-09-16
 - Improved description
 
-[Unreleased]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.2.2...HEAD
+[Unreleased]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.2.3...HEAD
+[1.2.3]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/imvs/spring-cloud-config-idea-plugin//compare/v1.0.4...v1.2.0